libfprint: Autospec creation for update from version 1.90.2 to version 1.90.3
2020-06-08: v1.90.3 release

This release mostly contains support for a number of new match-on-chip
devices. Most notable is the addition of the new goodixmoc driver.
Currently the driver has the small caveat that we have no strategy to
garbage collect old prints yet (a simple strategy could be implemented
in fprintd).

Highlights:
 * New goodixmoc driver supporting Goodix USB devices:
   27C6:5840
   27C6:6496
   27C6:60A2
 * Newly added support for Synaptics device:
   06CB:00E9

(NEWS truncated at 15 lines)